+++ packages/typeless/__tests__/unit/epic-hmr.test.tsx
@@ -1,9 +1,9 @@
 import * as React from 'react';
 import { act } from 'react-dom/test-utils';
-import { createModule } from '../src/createModule';
-import { startHmr, stopHmr } from '../src/onHmr';
-import { Registry } from '../src/Registry';
-import * as Rx from '../src/rx/rx';
+import { createModule } from '../../src/createModule';
+import { startHmr, stopHmr } from '../../src/onHmr';
+import { Registry } from '../../src/Registry';
+import * as Rx from '../../src/rx/rx';
 import { renderWithProvider } from './helpers';
 
 let container: HTMLDivElement = null!;
