+++ packages/typeless/__tests__/unit/epic-order.test.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import { act } from 'react-dom/test-utils';
-import { createModule } from '../src/createModule';
-import { Registry } from '../src/Registry';
+import { createModule } from '../../src/createModule';
+import { Registry } from '../../src/Registry';
 import { renderWithProvider } from './helpers';
 
 let container: HTMLDivElement = null!;
